move UserContext into hooks folder
